# docs/swagger.yaml
basePath: /
definitions:
  controller.Parfume:
    properties:
      _id:
        example: "123456789"
        type: string
      deskripsi:
        example: Parfum yang sangat wangi
        type: string
      harga:
        example: 1000000
        type: integer
      jenis_parfume:
        example: Eau de Parfum
        type: string
      merk:
        example: Dior
        type: string
      nama_parfume:
        example: Chirstian Dior
        type: string
      stok:
        example: 100
        type: integer
      tahun_peluncuran:
        example: 2021
        type: integer
      ukuran:
        example: 100ml
        type: string
    type: object
  controller.ReqParfume:
    properties:
      deskripsi:
        example: Parfum yang sangat wangi
        type: string
      harga:
        example: 1000000
        type: integer
      jenis_parfume:
        example: Eau de Parfum
        type: string
      merk:
        example: Dior
        type: string
      nama_parfume:
        example: Chirstian Dior
        type: string
      stok:
        example: 100
        type: integer
      tahun_peluncuran:
        example: 2021
        type: integer
      ukuran:
        example: 100ml
        type: string
    type: object
host: tb-parfume2024-34a7b650de40.herokuapp.com
info:
  contact:
    email: 714220013@std.ulbi.ac.id
    name: API Support
    url: https://github.com/barganakukuhraditya
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SWAGGER TUGAS BESAR
  version: "1.0"
paths:
  /delete/{id}:
    delete:
      consumes:
      - application/json
      description: Hapus data parfume.
      parameters:
      - description: Masukan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Delete data parfume.
      tags:
      - Parfume
  /insert:
    post:
      consumes:
      - application/json
      description: Input data parfume.
      parameters:
      - description: Payload Body [RAW]
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.Parfume'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Parfume'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Insert data parfume.
      tags:
      - Parfume
  /parfume:
    get:
      consumes:
      - application/json
      description: Mengambil semua data parfume.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Parfume'
      summary: Get All Data Parfume.
      tags:
      - Parfume
  /parfume/{id}:
    get:
      consumes:
      - application/json
      description: Ambil per ID data parfume.
      parameters:
      - description: Masukan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Parfume'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get By ID Data Parfume.
      tags:
      - Parfume
  /update/{id}:
    put:
      consumes:
      - application/json
      description: Ubah data parfume.
      parameters:
      - description: Masukan ID
        in: path
        name: id
        required: true
        type: string
      - description: Payload Body [RAW]
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.ReqParfume'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Parfume'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Update data parfume.
      tags:
      - Parfume
schemes:
- https
- http
swagger: "2.0"